requirements.txt
# ============================================
# FlashArb-Core Dependencies
# Python 3.10+ required
# ============================================

# --------------------------------------------
# Core: Web3 and Ethereum
# --------------------------------------------
web3>=6.0.0,<7.0.0
eth-abi>=5.0.0
eth-account>=0.11.0
eth-typing>=4.0.0
eth-utils>=4.0.0
hexbytes>=1.0.0

# --------------------------------------------
# Async HTTP Client
# --------------------------------------------
aiohttp>=3.9.0
aiohttp-socks>=0.8.0  # For proxy support (optional)

# --------------------------------------------
# Environment & Configuration
# --------------------------------------------
python-dotenv>=1.0.0
pydantic>=2.0.0  # For config validation (optional)

# --------------------------------------------
# Solidity Compiler (for contract deployment)
# --------------------------------------------
py-solc-x>=2.0.0

# --------------------------------------------
# Data Analysis & Visualization
# --------------------------------------------
pandas>=2.0.0
tabulate>=0.9.0
numpy>=1.24.0

# --------------------------------------------
# Logging & Monitoring
# --------------------------------------------
structlog>=23.0.0
rich>=13.0.0  # Beautiful terminal output

# --------------------------------------------
# Build Tools & Utilities
# --------------------------------------------
setuptools>=65.0.0
wheel>=0.42.0
typing-extensions>=4.0.0

# --------------------------------------------
# JSON & Data Validation
# --------------------------------------------
jsonschema>=4.0.0

# --------------------------------------------
# AWS Deployment (Optional - for cloud)
# --------------------------------------------
# boto3>=1.34.0  # AWS SDK (uncomment if needed)
# python-telegram-bot>=20.0  # Telegram alerts (optional)

# --------------------------------------------
# Development & Testing (Optional)
# --------------------------------------------
# pytest>=7.0.0
# pytest-asyncio>=0.21.0
# black>=23.0.0
# flake8>=6.0.0

